requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
ape-alchemy==0.8.0
ape-arbitrum==0.8.0
ape-ens==0.8.0
ape-etherscan==0.8.2
ape-foundry==0.8.0
ape-hardhat==0.8.0
ape-infura==0.8.0
ape-solidity==0.8.2
ape-template==0.8.0
ape-tokens==0.8.0
ape-vyper==0.8.3
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
base58==1.0.3
binaryornot==0.4.4
bitarray==2.9.2
cached-property==1.5.2
cbor2==5.6.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
chompjs==1.2.4
ckzg==1.0.2
click==8.1.7
cookiecutter==2.1.1
cytoolz==0.12.3
dataclassy==0.11.1
decorator==5.1.1
dnspython==2.6.1
eip712==0.2.7
email_validator==2.2.0
eth-account==0.11.2
eth-ape==0.8.9
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-pydantic-types==0.1.0
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.1.0
ethpm-types==0.6.14
evm-trace==0.2.0
evmchains==0.0.11
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ijson==3.3.0
importlib_metadata==8.0.0
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazyasd==0.1.4
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
morphys==1.0
msgspec==0.18.6
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.27.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==7.0.1
py-evm==0.10.1b1
py-geth==5.0.0b3
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.3
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.2.2
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rlp==4.0.1
rpds-py==0.19.0
safe-pysha3==1.0.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
text-unidecode==1.3
tokenlists==0.1.7
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
traitlets==5.14.3
trie==3.0.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
varint==1.0.2
vvm==0.2.1
vyper==0.3.10
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.13
web3==6.20.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
